feat: use separate provider for exported modals + fix when opening in veworld
